Changed Book Model, remigrated database, rewrote the default 400 response, initial EF repository code

+++ BooksApi/Migrations/DataContextModelSnapshot.cs
@@ -23,11 +23,11 @@ namespace BooksApi.Migrations
 
             modelBuilder.Entity("BooksApi.Models.Book", b =>
                 {
-                    b.Property<int>("Id")
+                    b.Property<long>("Id")
                         .ValueGeneratedOnAdd()
-                        .HasColumnType("int");
+                        .HasColumnType("bigint");
 
-                    SqlServerPropertyBuilderExtensions.UseIdentityColumn(b.Property<int>("Id"), 1L, 1);
+                    SqlServerPropertyBuilderExtensions.UseIdentityColumn(b.Property<long>("Id"), 1L, 1);
 
                     b.Property<string>("Author")
                         .IsRequired()
